Experimenting with new container and codec
Since OpenCV seems able to handle other containers now, we are moving away
from .avi (and it's small size limitation).

I experimented with 3 codecs for static and dynamic scenes and had mixed
results when comparing MPEG4 and H264; based on the Surface results (least
powerfull test device), I'm going with MPEG4 to begin with.
This choice is also better from the usability side since the user no
longer has to download the Cisco dlls.

MJPEG is faster, but the video size becomes huge too fast. In the future,
we make this parametrizable for the user similar to Pupil Labs.

On an additional note, the generated .mp4 videos are playable with Pupil Player
v0913.