Initial fisheye calibration working
WARNING: USE WITH CARE

It's not exactly trivial to calibrate the pupil labs wide angle with
the fisheye model, but this parameter set seems to do the trick.
The undistorted image looks acceptable, and the marker depth also seems
to check out (does the marker pose estimation takes the fisheye model into
account or is this just a coincidence?).

Calibrating that camera with a pinhole model yields acceptable undistorted
images, but the marker depth is very innacurate.

Further tests are pending.